Changed implementation to trigger on space.
As recommended in issue #1, we now trigger on space. We search backward to find the electric
character, and if it's present, we insert a newline. Inter-sentence and inter-word spacing is
handled by this code as well. Also, we no longer need to split abbreviations such as e.g. on the
separator characters.

Thus, this approach is a significant simplification of the code. It also works better, as it doesn't
trigger on references such as Fig.~1 (as there's no space).