Renamed electric regex variable, removed sentence-end from default value
To fix #10, the `sentence-end` function is included implicitly in the generated regex. This allows
customisations to sentence-end to have effect on twauctex functionality. However, it does not allow
users to remove the sentence-end value from twauctex-electric-regexes any more to use a fully custom
line ending. I would consider this niche functionality, however.